Separate files in my playlists page.
Separates styles file in my playlists.
